Move deleteRecord, mergeRecord into db utilities
