+++ src/sync/incomingSyncUtils.js
@@ -14,49 +14,7 @@ import { SETTINGS_KEYS } from '../settings';
 const { THIS_STORE_ID } = SETTINGS_KEYS;
 import { CHANGE_TYPES, generateUUID } from '../database';
 
-// Translations for merge logic.
-// TODO: bind translations to DataType constants to avoid breakage on schema update.
-const RECORD_TYPE_TO_TABLE = {
-  Item: {
-    StocktakeItem: {
-      field: 'item',
-    },
-    TransactionItem: {
-      field: 'item',
-    },
-    ItemBatch: {
-      field: 'item',
-    },
-    RequisitionItem: {
-      field: 'item',
-    },
-  },
-  Name: {
-    ItemBatch: {
-      field: 'supplier',
-    },
-    Transaction: {
-      field: 'otherParty',
-      setterMethod: 'setOtherParty',
-    },
-    Requisition: {
-      field: 'otherStoreName',
-    },
-  },
-};
-
-const RECORD_TYPE_TO_MASTERLIST = {
-  Item: {
-    MasterListItem: {
-      field: 'item',
-    },
-  },
-  Name: {
-    MasterListNameJoin: {
-      field: 'name',
-    },
-  },
-};
+import { deleteRecord, mergeRecords } from '../database/utilities';
 
 /**
  * Take the data from a sync record, and integrate it into the local database as
@@ -91,84 +49,6 @@ export function integrateRecord(database, settings, syncRecord) {
     default:
       return;
   }
-}
-
-/**
- * Merge two existing records. One record is retained and the other is merged. After the objects
- * are merged, the merged object is deleted from the database. The merge operation updates the
- * fields of the retained record to reference the same data as the merged object.
- *
- * @param {Realm}  database           The local database
- * @param {object} settings           Access to app settings
- * @param {string} internalRecordType Internal record type for merge operation
- * @param {object} syncRecord         Data representing the sync record
- */
-export function mergeRecords(database, settings, internalRecordType, syncRecord) {
-  const recordToKeep = database
-    .objects(internalRecordType)
-    .filtered('id == $0', syncRecord.mergeIDtokeep)[0];
-  const recordToMerge = database
-    .objects(internalRecordType)
-    .filtered('id == $0', syncRecord.mergeIDtodelete)[0];
-  const recordsExist = recordToKeep && recordToMerge;
-  if (!recordsExist) return;
-  const tablesToUpdate = RECORD_TYPE_TO_TABLE[internalRecordType];
-  if (!tablesToUpdate) return; // TODO: log to bugsnag if merging not implemented for a certain recordType.
-
-  Object.entries(tablesToUpdate).forEach(
-    ([tableToUpdate, { field: fieldToUpdate, setterMethod: fieldSetter }]) => {
-      const recordsToUpdate = database
-        .objects(tableToUpdate)
-        .filtered(`${fieldToUpdate}.id == $0`, recordToMerge.id)
-        .snapshot();
-      recordsToUpdate.forEach(record => {
-        if (record) {
-          if (typeof record[fieldSetter] === typeof Function) record[fieldSetter](recordToKeep);
-          else record[fieldToUpdate] = recordToKeep;
-        }
-      });
-    },
-  );
-
-  const [[tableToUpdate, { field: fieldToUpdate }]] = Object.entries(
-    RECORD_TYPE_TO_MASTERLIST[internalRecordType],
-  );
-  const masterListJoinRecords = database
-    .objects(tableToUpdate)
-    .filtered(`${fieldToUpdate}.id == $0`, recordToMerge.id)
-    .snapshot()
-    .forEach(joinRecord => {
-      const duplicateJoinRecord = database
-        .objects(tableToUpdate)
-        .filtered(
-          `(${fieldToUpdate}.id == $0) && (masterList.id == $0)`,
-          recordToKeep.id,
-          joinRecord.masterList.id,
-        )[0];
-      if (duplicateJoinRecord) {
-        deleteRecord(database, tableToUpdate, joinRecord.id);
-      } else {
-        joinRecord[fieldToUpdate] = recordToKeep;
-        createOrUpdateRecord(database, settings, tableToUpdate, joinRecord);
-      }
-    });
-
-  switch (internalRecordType) {
-    case 'Item':
-      recordToMerge.batches.forEach(batch => {
-        recordToKeep.addBatchIfUnique(batch);
-      });
-      break;
-    case 'Name':
-      recordToMerge.masterLists.forEach(masterList => {
-        recordToKeep.addMasterListIfUnique(masterList);
-      });
-      break;
-  }
-
-  recordToKeep.isVisible = recordToKeep.isVisible || recordToMerge.isVisible;
-
-  deleteRecord(database, internalRecordType, recordToMerge.id);
 }
 
 /**
@@ -536,67 +416,6 @@ export function createOrUpdateRecord(database, settings, recordType, record) {
       database.save('Transaction', transaction);
       itemBatch.addTransactionBatchIfUnique(transactionBatch);
       database.save('ItemBatch', itemBatch);
-      break;
-    }
-    default:
-      break; // Silently ignore record types we don't want to sync into mobile
-  }
-}
-
-/**
- * Delete the record with the given id, relying on its destructor to initiate any
- * changes that are required to clean up that type of record.
- * @param  {Realm}  database   App wide local database
- * @param  {string} recordType Internal record type
- * @param  {string} primaryKey       The primary key of the database object, usually its id
- * @param  {string} primaryKeyField  The field used as the primary key, defaults to 'id'
- * @return {none}
- */
-function deleteRecord(database, recordType, primaryKey, primaryKeyField = 'id') {
-  // 'delete' is a reserved word, deleteRecord is in the upper scope, so here we have:
-  const obliterate = () => {
-    const deleteResults = database
-      .objects(recordType)
-      .filtered(`${primaryKeyField} == $0`, primaryKey);
-    if (deleteResults && deleteResults.length > 0) database.delete(recordType, deleteResults[0]);
-  };
-
-  switch (recordType) {
-    case 'Item':
-    case 'ItemBatch':
-    case 'ItemCategory':
-    case 'ItemDepartment':
-    case 'ItemStoreJoin':
-    case 'MasterList':
-    case 'MasterListItem':
-    case 'Name':
-    case 'NameStoreJoin':
-    case 'NumberSequence':
-    case 'NumberToReuse':
-    case 'Requisition':
-    case 'RequisitionItem':
-    case 'Stocktake':
-    case 'StocktakeBatch':
-    case 'Transaction':
-    case 'TransactionBatch':
-    case 'TransactionCategory': {
-      obliterate();
-      break;
-    }
-    // LocalListItem is mimicked with MasterListItem
-    case 'LocalListItem':
-      deleteRecord(database, 'MasterListItem', primaryKey, primaryKeyField);
-      break;
-    case 'MasterListNameJoin': {
-      // Joins for local lists are mapped to and mimicked by a MasterList of the same id.
-      const masterList = database.objects('MasterList').filtered('id == $0', primaryKey)[0];
-      if (masterList) {
-        // Is a local list, so delete the MasterList that was created for it.
-        deleteRecord(database, 'MasterList', primaryKey, primaryKeyField);
-      } else {
-        // Delete the MasterListNameJoin as in the normal/expected case.
-        obliterate();
-      }
       break;
     }
     default:
